requirements.txt
flask==2.0.1
werkzeug==2.0.1
flask-cors==3.0.10
openai==1.8.0
assemblyai==0.17.0
python-dotenv==1.0.0
pydub==0.25.1
psutil==5.9.8